% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-compounds.R
\name{.create_single_compound}
\alias{.create_single_compound}
\title{Creates a single compound}
\usage{
.create_single_compound(compound, wordsep = " ", concatenator = "_")
}
\arguments{
\item{compound}{A simple text (= character vector).}

\item{wordsep}{The word separator to look for, usually spaces. Defaults to
" ".}

\item{concatenator}{The replacement for wordsep characters (e.g., " ").
Defaults to "_".}
}
\value{
The provided text where all occurrences of the wordsep character
(by default, " ") were replaced by the concatenator character (by default,
"_").
}
\description{
This is an internal function use in the
.make_compounds_for_single_pattern function. It is basically a wrapper around
stringr::str_replace_all. By default, it replaces all space characters (" ")
with underline character ("_").
}
