% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-compounds.R
\name{.make_compounds_for_single_pattern}
\alias{.make_compounds_for_single_pattern}
\title{Replace matches for a single pattern with compounds}
\usage{
.make_compounds_for_single_pattern(
  text,
  pattern,
  wordsep = " ",
  concatenator = "_",
  lazy = TRUE,
  ignore_case = TRUE
)
}
\arguments{
\item{text}{Text should be provided as a character vector.}

\item{pattern}{A single pattern with possible regular expressions.}

\item{wordsep}{The word separator to look for, usually spaces. Defaults to
" ".}

\item{concatenator}{The replacement for wordsep characters (e.g., " ").
Defaults to "_".}

\item{lazy}{Should regular expressions be transformed to lazy versions?
Defaults to TRUE to return the shortest possible compounds.}

\item{ignore_case}{Should the case be ignored when searching for pattern
matches? Defaults to TRUE.}
}
\value{
The provided text where all occurrences of the pattern have been
converted to multi-word compounds.
}
\description{
This is an internal function use in the make_compounds function.
}
