% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-util-functions.R
\name{every_term_a_dict}
\alias{every_term_a_dict}
\title{Make every term a dictionary}
\usage{
every_term_a_dict(dict)
}
\arguments{
\item{dict}{A quanteda dictionary or a simple list of dictionary terms}
}
\value{
A quanteda dictionary object where each term is its own dictionary.
Dictionary names are simply the provided terms.
}
\description{
This functions transforms a dictionary (or a list of terms) to
a quanteda dictionary object where each term is it's own dictionary.
}
