% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-util-functions.R
\name{run_kwic_stepwise}
\alias{run_kwic_stepwise}
\title{Stepwise locate keywords-in-context}
\usage{
run_kwic_stepwise(
  tokens,
  pattern,
  window = 25,
  valuetype = "regex",
  step = 10000,
  ...
)
}
\arguments{
\item{tokens}{A quanteda tokens object.}

\item{pattern}{A character vector, list of character vectors, dictionary, or
collocations object.}

\item{window}{The number of context words to be displayed around the keyword.}

\item{valuetype}{The type of pattern matching: "glob" for "glob"-style
wildcard expressions; "regex" for regular expressions; or "fixed" for exact
matching. Defaults to "regex".}

\item{step}{How many tokens should be processed at once? Defaults to 10000.}

\item{...}{Additonal arguments passed to the kwic function.}
}
\value{
A kwic classed data.frame.
}
\description{
With large amounts of text and complex (regex) patterns you
might run into problems with memory when running quanteda's kwic function.
Thus this function takes a stepwise approach. The tokens are split into
certain chunks (specified by step) and then kwic is run on these smaller
chunks.
}
